Split destination classes into their own files
