+++ importer/Importers/destinations/Elkarte1_0/ImporterStep3.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * @name      OpenImporter
+ * @copyright OpenImporter contributors
+ * @license   BSD http://opensource.org/licenses/BSD-3-Clause
+ *
+ * @version 2.0 Alpha
+ *
+ * This file contains code based on:
+ *
+ * Simple Machines Forum (SMF)
+ * copyright:	2011 Simple Machines (http://www.simplemachines.org)
+ * license:	BSD, See included LICENSE.TXT for terms and conditions.
+ */
+
+namespace OpenImporter\Importers\destinations\ElkArte1_0;
+
+class ImporterStep3 extends \OpenImporter\Importers\SmfCommonOriginStep3
+{
+}
